refactor: Service 객체 역할을 Controller 메소드로 변경
